test: move spec_decode_verify test into python/tests/test_fast.py
The spec_decode_verify validation lived in a root-level test_spec_verify.py
script that pytest does not collect, so it never ran in CI. Fold its coverage
into TestFast.test_spec_decode_verify (oracle-checked cases across
fp32/fp16/bf16, both the fused-kernel and cpu-composition paths, a randomized
sweep, and the bad-shape rejection) and drop the root script.
